Add additional unit tests for new code refactor
