+++ test/common/schemaform/containers/FormApp.unit.spec.jsx
@@ -28,77 +28,26 @@ describe('Schemaform <FormApp>', () => {
     expect(tree.everySubTree('FormNav')).to.be.empty;
     expect(tree.everySubTree('FormTitle')).to.be.empty;
   });
-  // it('should not render form title or nav on intro page when disableSave is true', () => {
-  //   const formConfig = {
-  //     disableSave: true
-  //   };
-  //   const currentLocation = {
-  //     pathname: 'introduction',
-  //     search: ''
-  //   };
-  //   const routes = [{
-  //     pageList: [{ path: currentLocation.pathname }]
-  //   }];
+  it('should show nav when the form is in progress', () => {
+    const formConfig = {};
+    const currentLocation = {
+      pathname: '/veteran-information/personal-information',
+      search: ''
+    };
+    const routes = [{
+      pageList: [{ path: currentLocation.pathname }]
+    }];
 
-  //   const tree = SkinDeep.shallowRender(
-  //     <FormApp
-  //       formConfig={formConfig}
-  //       routes={routes}
-  //       currentLocation={currentLocation}>
-  //       <div className="child"/>
-  //     </FormApp>
-  //   );
+    const tree = SkinDeep.shallowRender(
+      <FormApp
+        formConfig={formConfig}
+        routes={routes}
+        currentLocation={currentLocation}>
+        <div className="child"/>
+      </FormApp>
+    );
 
-  //   expect(tree.everySubTree('.child')).not.to.be.empty;
-  //   expect(tree.everySubTree('FormNav')).to.be.empty;
-  //   expect(tree.everySubTree('FormTitle')).to.be.empty;
-  // });
-  // it('should not render form title and nav when save is enabled', () => {
-  //   const formConfig = {};
-  //   const currentLocation = {
-  //     pathname: 'test',
-  //     search: ''
-  //   };
-  //   const routes = [{
-  //     pageList: [{ path: currentLocation.pathname }]
-  //   }];
-
-  //   const tree = SkinDeep.shallowRender(
-  //     <FormApp
-  //       formConfig={formConfig}
-  //       routes={routes}
-  //       currentLocation={currentLocation}>
-  //       <div className="child"/>
-  //     </FormApp>
-  //   );
-
-  //   expect(tree.everySubTree('.child')).not.to.be.empty;
-  //   expect(tree.everySubTree('FormTitle')).to.be.empty;
-  //   expect(tree.everySubTree('FormNav')).to.be.empty;
-  // });
-  // it('should render form title and nav when save is disabled', () => {
-  //   const formConfig = {
-  //     title: 'Testing',
-  //     disableSave: true
-  //   };
-  //   const currentLocation = {
-  //     pathname: 'test',
-  //     search: ''
-  //   };
-  //   const routes = [{
-  //     pageList: [{ path: currentLocation.pathname }]
-  //   }];
-
-  //   const tree = SkinDeep.shallowRender(
-  //     <FormApp
-  //       formConfig={formConfig}
-  //       routes={routes}
-  //       currentLocation={currentLocation}>
-  //       <div className="child"/>
-  //     </FormApp>
-  //   );
-
-  //   expect(tree.everySubTree('FormTitle')).not.to.be.empty;
-  //   expect(tree.everySubTree('FormNav')).not.to.be.empty;
-  // });
+    expect(tree.everySubTree('.child')).not.to.be.empty;
+    expect(tree.everySubTree('FormNav')).not.to.be.empty;
+  });
 });
