+++ test/common/schemaform/save-in-progress/RoutedSavableApp.unit.spec.jsx
@@ -52,53 +52,6 @@ describe('Schemaform <RoutedSavableApp>', () => {
     expect(tree.everySubTree('FormNav')).to.be.empty;
     expect(tree.everySubTree('FormTitle')).to.be.empty;
   });
-  // it('should render nav and children', () => {
-  //   const formConfig = {};
-  //   const currentLocation = {
-  //     pathname: 'test',
-  //     search: ''
-  //   };
-  //   const routes = [{
-  //     pageList: [{ path: currentLocation.pathname }]
-  //   }];
-
-  //   const tree = SkinDeep.shallowRender(
-  //     <RoutedSavableApp
-  //       formConfig={formConfig}
-  //       routes={routes}
-  //       currentLocation={currentLocation}
-  //       loadedStatus={LOAD_STATUSES.notAttempted}>
-  //       <div className="child"/>
-  //     </RoutedSavableApp>
-  //   );
-
-  //   expect(tree.everySubTree('.child')).not.to.be.empty;
-  //   expect(tree.everySubTree('FormNav')).not.to.be.empty;
-  // });
-  // it('should render form title', () => {
-  //   const formConfig = {
-  //     title: 'Testing'
-  //   };
-  //   const currentLocation = {
-  //     pathname: 'test',
-  //     search: ''
-  //   };
-  //   const routes = [{
-  //     pageList: [{ path: currentLocation.pathname }]
-  //   }];
-
-  //   const tree = SkinDeep.shallowRender(
-  //     <RoutedSavableApp
-  //       formConfig={formConfig}
-  //       routes={routes}
-  //       currentLocation={currentLocation}
-  //       loadedStatus={LOAD_STATUSES.notAttempted}>
-  //       <div className="child"/>
-  //     </RoutedSavableApp>
-  //   );
-
-  //   expect(tree.everySubTree('FormTitle')).not.to.be.empty;
-  // });
   it('should render the loading screen', () => {
     const formConfig = {
       title: 'Testing'
